[raw2pgm] Shell script for converting Bayer RAW files to Bayer PGM
The conversion is not touching the pixel data, just adding a PGM header to it.
This is nonetheless a useful operation, because compared to headerless RAW,
PGM is better supported by third-party tools and easier to work with. Image
resolution and bit depth are hardcoded.

Depends on 'rawtopgm' from Netpbm; 'sudo apt install netpbm' to get it.